Data.Array.Accelerate.Data.Colour.HSL

Description

Colours in the HSL (hue-saturation-lightness) cylindrical-coordinate representation of points in the RGB colour space.

Documentation

type Colour = HSL Float

A HSL colour value

data HSL a

Constructors

HSL a a a

hsl

Arguments

:: Exp Float
hue component

-> Exp Float
saturation component

-> Exp Float
lightness component

-> Exp Colour

Construct a HSL colour value from the individual channel components. The hue component is measured in degrees and wrapped to the range [0..360), while the saturation and value are clamped to the range [0..1].

clamp :: Exp Colour -> Exp Colour

toRGB :: Exp (HSL Float) -> Exp (RGB Float)

Convert a HSL colour to an RGB colour-space value

fromRGB :: Exp (RGB Float) -> Exp (HSL Float)

Convert a point in the RGB colour-space to a point in the HSL colour-space.

hue :: Exp (RGB Float) -> Exp Float

Return the HSL-hue of an RGB colour

saturation :: Exp (RGB Float) -> Exp Float

Return the HSL-saturation of an RGB colour

lightness :: Exp (RGB Float) -> Exp Float

Return the HSL-lightness of an RGB colour