Data.Array.Accelerate.Data.Colour.RGB

Description

RGB triples for an unspecified colour space

Documentation

type Colour = RGB Float

An RGB colour value

data RGB a

Constructors

RGB a a a

rgb

Arguments

:: Exp Float
red component

-> Exp Float
green component

-> Exp Float
blue component

-> Exp Colour

Construct an RGB colour from individual channel components. The components will be clamped to the range [0..1].

rgb8

Arguments

:: Exp Word8
red component

-> Exp Word8
green component

-> Exp Word8
blue component

-> Exp Colour

Construct a colour from 8-bit-per-channel colour components.

clamp :: Exp Colour -> Exp Colour

Clamp each component of a colour to the range [0..1].

blend

Arguments

:: Exp Float
Proportion of first colour

-> Exp Float
Proportion of second colour

-> Exp Colour
First colour

-> Exp Colour
Second colour

-> Exp Colour
Resulting colour

Blend two colours in the given proportions.

Note that this uses an approximation of gamma=2 (i.e. sum-of-squares method). It is recommended to instead convert to the sRGB colour space if you want more accurate colour blending, or if you intend to use the gamma-corrected values more than once (e.g. in a stencil).

blend c1 c2 ~= SRGB.toRGB ( (SRGB.fromRGB c1 + SRGB.fromRGB c2) / 2 )

See the Blur program in the examples for a comparison of mixing colours in the RGB and sRGB colour spaces.

luminance :: Exp Colour -> Exp Float

Luminance of an RGB colour (Y component of a YUV colour).

packRGB :: Exp Colour -> Exp Word32

Convert a Colour into a packed-word RGBA representation

packBGR :: Exp Colour -> Exp Word32

Convert a colour into a packed-word ABGR representation

unpackRGB :: Exp Word32 -> Exp Colour

Convert a colour from a packed-word RGBA representation

unpackBGR :: Exp Word32 -> Exp Colour

Convert a colour from a packed-word ABGR representation

packRGB8 :: Exp (RGB Word8) -> Exp Word32

packBGR8 :: Exp (RGB Word8) -> Exp Word32

unpackRGB8 :: Exp Word32 -> Exp (RGB Word8)

unpackBGR8 :: Exp Word32 -> Exp (RGB Word8)