Data.Array.Accelerate.Data.Colour.RGBA

Description

RGBA quadruples for an unspecified colour space

Documentation

type Colour = RGBA Float

An RGBA colour value.

data RGBA a

An RGBA colour value to hold the colour components. All components lie in the range [0..1).

We need to parameterise by a type so that we can have both Exp (RGBA a) and RGBA (Exp a).

Constructors

RGBA a a a a

rgba

Arguments

:: Exp Float
red component

-> Exp Float
green component

-> Exp Float
blue component

-> Exp Float
alpha component

-> Exp Colour

Construct an RGBA colour from individual channel components. The components will be clamped to the range [0..1].

rgba8

Arguments

:: Exp Word8
red component

-> Exp Word8
green component

-> Exp Word8
blue component

-> Exp Word8
alpha component

-> Exp Colour

Construct a colour from 8-bits-per-channel colour components.

clamp :: Exp Colour -> Exp Colour

Clamp each component to the range [0..1]

blend

Arguments

:: Exp Float
Proportion of first colour

-> Exp Float
Proportion of second colour

-> Exp Colour
First colour

-> Exp Colour
Second colour

-> Exp Colour
Resulting colour

Blend two colours in the given proportions.

Note that this uses an approximation of gamma=2 (i.e. sum-of-squares method). It is recommended to instead convert to the sRGB colour space if you want more accurate colour blending, or if you intend to use the gamma-corrected values more than once (e.g. in a stencil).

blend c1 c2 ~= SRGB.toRGB ( (SRGB.fromRGB c1 + SRGB.fromRGB c2) / 2 )

See the Blur program in the examples for a comparison of mixing colours in the RGB and sRGB colour spaces.

luminance :: Exp Colour -> Exp Float

Luminance of an RGB colour (Y component of a YUV colour).

opacity :: Exp Float -> Exp Colour -> Exp Colour

Set the opacity of the given colour. The opacity is clamped to the range [0..1].

opaque :: Exp Colour -> Exp Colour

A completely opaque colour

transparent :: Exp Colour -> Exp Colour

Make colour transparent

packRGBA :: Exp Colour -> Exp Word32

Convert a Colour into a packed-word RGBA representation

packABGR :: Exp Colour -> Exp Word32

Convert a colour into a packed-word ABGR representation

unpackRGBA :: Exp Word32 -> Exp Colour

Convert a colour from a packed-word RGBA representation

unpackABGR :: Exp Word32 -> Exp Colour

Convert a colour from a packed-word ABGR representation

packRGBA8 :: Exp (RGBA Word8) -> Exp Word32

packABGR8 :: Exp (RGBA Word8) -> Exp Word32

unpackRGBA8 :: Exp Word32 -> Exp (RGBA Word8)

unpackABGR8 :: Exp Word32 -> Exp (RGBA Word8)