Data.Array.Accelerate.Data.Colour.HSV

Description

Colours in the HSV (hue-saturation-value) cylindrical-coordinate representation of points in the RGB colour space.

Documentation

type Colour = HSV Float

A HSV colour value

data HSV a

Constructors

HSV a a a

hsv

Arguments

:: Exp Float
hue component

-> Exp Float
saturation component

-> Exp Float
value component

-> Exp Colour

Construct a HSV colour value from the individual channel components. The hue component is measured in degrees and wrapped to the range [0..360), while the saturation and value are clamped to the range [0..1].

clamp :: Exp Colour -> Exp Colour

toRGB :: Exp (HSV Float) -> Exp (RGB Float)

Convert a HSV colour to an RGB colour-space value

fromRGB :: Exp (RGB Float) -> Exp (HSV Float)

Convert a point in the RGB colour-space to a point in the HSV colour-space.

hue :: Exp (RGB Float) -> Exp Float

Return the HSV-hue of an RGB colour

saturation :: Exp (RGB Float) -> Exp Float

Return the HSV-saturation of an RGB colour

value :: Exp (RGB Float) -> Exp Float

Return the HSV-value of an RGB colour